modules/llm_app/src/requirements.txt
azure-search-documents==11.6.0b4
azure-storage-blob
azure-identity
azure-mgmt-resource
azure-mgmt-web
azure-mgmt-storage